macOS libc headers: downgrade to 10.15.7 (Catalina)
See ziglang/fetch-them-macos-headers#2 for more details. The path
forward looks like one of the following:

 * Ony provide headers for the oldest supported macOS (currently 10.13
   but soon to be bumped to 10.14).
 * Provide headers for multiple versions, and select based on the Zig
   target OS version range minimum.
 * Don't try to provide macOS headers.

If we don't tackle the version problem, we would have to re-introduce
the ability to detect and depend on native system headers if we wanted
to support C/C++ code that used newer OS definitions.

This patch also adds support for `#include <mach/mach_time.h>`.

Also related: #5236

+++ lib/libc/include/x86_64-macos-gnu/dispatch/queue.h
@@ -661,7 +661,7 @@ typedef long dispatch_queue_priority_t;
 API_AVAILABLE(macos(10.6), ios(4.0))
 DISPATCH_EXPORT DISPATCH_CONST DISPATCH_WARN_RESULT DISPATCH_NOTHROW
 dispatch_queue_global_t
-dispatch_get_global_queue(intptr_t identifier, uintptr_t flags);
+dispatch_get_global_queue(long identifier, unsigned long flags);
 
 /*!
  * @typedef dispatch_queue_attr_t
